Inline ckbtc_minter_basics_test
